rename SamplePoint to Spacetime and move them one dir up
