+++ tests/instruments/test_adcp.py
@@ -3,7 +3,7 @@
 import numpy as np
 from parcels import FieldSet
 
-from virtual_ship.instruments import Location, Spacetime
+from virtual_ship import Location, Spacetime
 from virtual_ship.instruments.adcp import simulate_adcp
