+++ tests/instruments/test_argo.py
@@ -5,7 +5,7 @@ from datetime import timedelta
 import numpy as np
 from parcels import FieldSet
 
-from virtual_ship.instruments import Location
+from virtual_ship import Location, Spacetime
 from virtual_ship.instruments.argo_float import ArgoFloat, simulate_argo_floats
 
 
@@ -29,8 +29,7 @@ def test_simulate_argo_floats() -> None:
 
     argo_floats = [
         ArgoFloat(
-            location=Location(latitude=0, longitude=0),
-            deployment_time=0,
+            spacetime=Spacetime(location=Location(latitude=0, longitude=0), time=0),
             min_depth=min_depth,
             max_depth=MAX_DEPTH,
             drift_depth=DRIFT_DEPTH,
