+++ virtual_ship/__init__.py
@@ -1,5 +1,7 @@
 """Code for the Virtual Ship Classroom, where Marine Scientists can combine Copernicus Marine Data with an OceanParcels ship to go on a virtual expedition."""
 
 from . import instruments, sailship
+from .location import Location
+from .spacetime import Spacetime
 
-__all__ = ["instruments", "sailship"]
+__all__ = ["Location", "Spacetime", "instruments", "sailship"]
