+++ virtual_ship/instruments/__init__.py
@@ -1,7 +1,5 @@
 """Measurement instrument that can be used with Parcels."""
 
-from . import argo_float, ctd, drifter
-from .location import Location
-from .spacetime import Spacetime
+from . import adcp, argo_float, ctd, drifter, ship_st
 
-__all__ = ["Location", "Spacetime", "argo_float", "ctd", "drifter"]
+__all__ = ["adcp", "argo_float", "ctd", "drifter", "ship_st"]
